refactor: move JSON array and JSON object to common module
