+++ AnkiDroid/src/main/java/com/ichi2/anki/FilteredDeckOptions.kt
@@ -28,13 +28,13 @@ import android.preference.Preference
 import android.preference.PreferenceCategory
 import androidx.core.content.edit
 import com.ichi2.anki.analytics.UsageAnalytics
+import com.ichi2.anki.common.utils.ext.stringIterable
 import com.ichi2.annotations.NeedsTest
 import com.ichi2.libanki.Collection
 import com.ichi2.preferences.StepsPreference.Companion.convertFromJSON
 import com.ichi2.preferences.StepsPreference.Companion.convertToJSON
 import com.ichi2.themes.Themes
 import com.ichi2.ui.AppCompatPreferenceActivity
-import com.ichi2.utils.stringIterable
 import org.json.JSONArray
 import org.json.JSONException
 import org.json.JSONObject
