+++ AnkiDroid/src/main/java/com/ichi2/libanki/CardTemplate.kt
@@ -16,6 +16,7 @@
 
 package com.ichi2.libanki
 
+import com.ichi2.anki.common.utils.ext.deepClone
 import com.ichi2.utils.JSONObjectHolder
 import com.ichi2.utils.deepClone
 import net.ankiweb.rsdroid.RustCleanup
