+++ AnkiDroid/src/main/java/com/ichi2/libanki/NotetypeJson.kt
@@ -17,8 +17,8 @@
 package com.ichi2.libanki
 
 import androidx.annotation.CheckResult
+import com.ichi2.anki.common.utils.ext.toStringList
 import com.ichi2.utils.deepClonedInto
-import com.ichi2.utils.toStringList
 import org.intellij.lang.annotations.Language
 import org.json.JSONArray
 import org.json.JSONObject
