+++ AnkiDroid/src/main/java/com/ichi2/preferences/StepsPreference.kt
@@ -22,8 +22,8 @@ import android.util.AttributeSet
 import android.view.View
 import com.ichi2.anki.AnkiDroidApp
 import com.ichi2.anki.R
+import com.ichi2.anki.common.utils.ext.stringIterable
 import com.ichi2.anki.showThemedToast
-import com.ichi2.utils.stringIterable
 import org.json.JSONArray
 import org.json.JSONException
 import timber.log.Timber
