+++ AnkiDroid/src/main/java/com/ichi2/utils/JSONContainer.kt
@@ -17,6 +17,7 @@
 package com.ichi2.utils
 
 import androidx.annotation.VisibleForTesting
+import com.ichi2.anki.common.utils.ext.jsonObjectIterator
 import com.ichi2.libanki.utils.NotInLibAnki
 import com.ichi2.libanki.utils.append
 import com.ichi2.libanki.utils.index
