+++ AnkiDroid/src/test/java/com/ichi2/libanki/PythonExtensionsTest.kt
@@ -17,8 +17,8 @@
 package com.ichi2.libanki
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.ichi2.anki.common.utils.ext.jsonObjectIterable
 import com.ichi2.libanki.utils.insert
-import com.ichi2.utils.jsonObjectIterable
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.collection.IsIterableContainingInOrder.contains
